Avoid code duplication between the 2 tests.
